+++ src/components/Label/LabelTable/Body/LabelRow.jsx
@@ -4,10 +4,10 @@ import {
   LabelUpdateFormWrap,
   RowInfo
 } from './LabelRow.style'
-import { ALERT_MESSAGE, BTN_LABEL } from '../../../utils/constant'
+import { ALERT_MESSAGE, BTN_LABEL } from '../../../../utils/constant'
 import LabelForm from '../../LabelForm/LabelForm'
 import { useState } from 'react'
-import { deleteLabel } from '../../../utils/api'
+import { deleteLabel } from '../../../../utils/api'
 
 const LabelRow = ({ label, updateData }) => {
   const { id = null, labelName = '', desc = '', color = '#e99695' } = label
