+++ src/components/Milestones/CreateForm/CreateForm.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const CreateForm = () => {
+  return (
+    <div>
+      MilesStone
+    </div>
+  )
+}
+
+export default CreateForm
