+++ src/components/Milestones/Milestones.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Milestones = () => {
+  return (
+    <div>
+      Milestones
+    </div>
+  )
+}
+
+export default Milestones
