+++ src/components/Milestones/Table/Table.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Table = () => {
+  return (
+    <div>
+      Milestone
+    </div>
+  )
+}
+
+export default Table
